Revert "oobmigration: Relocate insights migration code (#40580)"
This reverts commit ed1e2309b9526cba17c3e54e09b9414f4623f477.
