+++ enterprise/internal/insights/migration/discovery.go
@@ -1,4 +1,6 @@
-package insights
+// Re-purposed and copied methods from discovery and other related methods.
+
+package migration
 
 import (
 	"context"
