+++ enterprise/internal/insights/migration/migration.go
@@ -1,16 +1,13 @@
-package insights
+package migration
 
 import (
 	"context"
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/inconshreveable/log15"
 
 	"github.com/keegancsmith/sqlf"
-
-	"github.com/sourcegraph/log"
 
 	edb "github.com/sourcegraph/sourcegraph/enterprise/internal/database"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/store"
@@ -18,7 +15,9 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/insights"
+	"github.com/sourcegraph/sourcegraph/internal/oobmigration"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -30,6 +29,7 @@ const (
 )
 
 type migrator struct {
+	insightsDB dbutil.DB
 	postgresDB database.DB
 
 	settingsMigrationJobsStore *store.DBSettingsMigrationJobsStore
@@ -40,23 +40,18 @@ type migrator struct {
 	workerBaseStore            *basestore.Store
 }
 
-func NewMigrator(frontendStore, insightsStore *basestore.Store) *migrator {
-	logger := log.Scoped("insights-migration", "")
-	postgresDB := database.NewDBWith(logger, frontendStore)
-
+func NewMigrator(insightsDB edb.InsightsDB, postgresDB database.DB) oobmigration.Migrator {
 	return &migrator{
+		insightsDB:                 insightsDB,
 		postgresDB:                 postgresDB,
 		settingsMigrationJobsStore: store.NewSettingsMigrationJobsStore(postgresDB),
 		settingsStore:              postgresDB.Settings(),
-		insightStore:               &store.InsightStore{Store: insightsStore, Now: time.Now},
-		dashboardStore:             &store.DBDashboardStore{Store: insightsStore, Now: time.Now},
+		insightStore:               store.NewInsightStore(insightsDB),
+		dashboardStore:             store.NewDashboardStore(insightsDB),
 		orgStore:                   postgresDB.Orgs(),
 		workerBaseStore:            basestore.NewWithHandle(postgresDB.Handle()),
 	}
 }
-
-func (m *migrator) ID() int                 { return 14 }
-func (m *migrator) Interval() time.Duration { return time.Second * 10 }
 
 func (m *migrator) Progress(ctx context.Context) (float64, error) {
 	progress, _, err := basestore.ScanFirstFloat(m.settingsMigrationJobsStore.Query(ctx, sqlf.Sprintf(`
