+++ enterprise/internal/insights/migration/migration_no_op.go
@@ -1,4 +1,4 @@
-package insights
+package migration
 
 import (
 	"context"
