+++ enterprise/internal/insights/migration/migration_test.go
@@ -1,4 +1,4 @@
-package insights
+package migration
 
 import (
 	"context"
